fix(cache): clone `Response` before `Set-Cookie` guard;
This fixes a bug with worktop that *only* occurred when a handler returned a fetch() call directly and the subrequest was to a resource that was cacheable & had a "Set-Cookie" header.

In code, this may have looked like:

```js
API.add('GET', '/example', (req, res) => {
  return fetch('https://other.com/foobar');
});
```

And where the "https://other.com/foobar" response was cacheable and had a "Set-Cookie" header.

Any/All other cache usage was (and remains) unaffected.